+++ Model/ReaderWriter.php
@@ -1,0 +1,198 @@
+<?php
+
+namespace Nuvei\Checkout\Model;
+
+/**
+ * Read and Write onto the file system.
+ *
+ * @author Nuvei
+ */
+class ReaderWriter
+{
+    private $fileSystem;
+    private $config;
+    private $directory;
+    private $traceId;
+    
+    public function __construct(
+        \Magento\Framework\ObjectManagerInterface $objectManager
+        ,\Nuvei\Checkout\Model\Config $config
+        ,\Magento\Framework\Filesystem\DirectoryList $directory
+    ) {
+        try {
+            $this->fileSystem   = $objectManager->create(\Magento\Framework\Filesystem\DriverInterface::class);
+            $this->config       = $config;
+            $this->directory    = $directory;
+        } catch(Exception $ex) {
+            $this->createLog($ex->getMessage(), 'ReaderWriter Exception');
+        }
+    }
+    
+    /**
+     * Prepare and save log.
+     * 
+     * @param mixed $data
+     * @param string $title
+     * @param string $log_level
+     * 
+     * @return void
+     */
+    public function createLog($data, $title = '', $log_level = 'TRACE')
+    {
+        if (! $this->config->isDebugEnabled()) {
+            return;
+        }
+        
+        $logsPath   = $this->directory->getPath('log');
+        $d          = $data;
+        $string     = '';
+        
+        if (is_bool($data)) {
+            $d = $data ? 'true' : 'false';
+        } elseif (is_string($data) || is_numeric($data)) {
+            $d = $data;
+        } elseif ('' === $data) {
+            $d = 'Data is Empty.';
+        } elseif (is_array($data)) {
+            // do not log accounts if on prod
+            if (!$this->config->isTestModeEnabled()) {
+                if (isset($data['userAccountDetails']) && is_array($data['userAccountDetails'])) {
+                    $data['userAccountDetails'] = 'account details';
+                }
+                if (isset($data['userPaymentOption']) && is_array($data['userPaymentOption'])) {
+                    $data['userPaymentOption'] = 'user payment options details';
+                }
+                if (isset($data['paymentOption']) && is_array($data['paymentOption'])) {
+                    $data['paymentOption'] = 'payment options details';
+                }
+            }
+            // do not log accounts if on prod
+
+            if (!empty($data['paymentMethods']) && is_array($data['paymentMethods'])) {
+                $data['paymentMethods'] = json_encode($data['paymentMethods']);
+            }
+            if (!empty($data['Response data']['paymentMethods'])
+                && is_array($data['Response data']['paymentMethods'])
+            ) {
+                $data['Response data']['paymentMethods'] = json_encode($data['Response data']['paymentMethods']);
+            }
+
+            if (!empty($data['plans']) && is_array($data['plans'])) {
+                $data['plans'] = json_encode($data['plans']);
+            }
+
+            $d = $this->config->isTestModeEnabled() ? json_encode($data, JSON_PRETTY_PRINT) : json_encode($data);
+        } elseif (is_object($data)) {
+            $d = $this->config->isTestModeEnabled() ? json_encode($data, JSON_PRETTY_PRINT) : json_encode($data);
+        } else {
+            $d = $this->config->isTestModeEnabled() ? json_encode($data, JSON_PRETTY_PRINT) : json_encode($data);
+        }
+        
+        $tab            = '    ';
+        $utimestamp     = microtime(true);
+        $timestamp      = floor($utimestamp);
+        $milliseconds   = round(($utimestamp - $timestamp) * 1000000);
+        $record_time    = date('Y-m-d') . 'T' . date('H:i:s') . '.' . $milliseconds . date('P');
+        
+        if(!$this->traceId) {
+            $this->traceId = bin2hex(random_bytes(16));
+        }
+        
+        $source_file_name   = '';
+        $member_name        = '';
+        $source_line_number = '';
+        
+        $backtrace = debug_backtrace();
+        if(!empty($backtrace)) {
+            if(!empty($backtrace[0]['file'])) {
+                $file_path_arr  = explode(DS, $backtrace[0]['file']);
+                
+                if(!empty($file_path_arr)) {
+                    $source_file_name = end($file_path_arr) . '|';
+                }
+            }
+            
+//            if(!empty($backtrace[0]['function'])) {
+//                $member_name = $backtrace[0]['function'] . '|';
+//            }
+            
+            if(!empty($backtrace[0]['line'])) {
+                $source_line_number = $backtrace[0]['line'] . $tab;
+            }
+        }
+        
+        $string .= $record_time . $tab
+            . $log_level . $tab
+            . $this->traceId . $tab
+            . 'Checkout ' . $this->config->getSourcePlatformField() . '|'
+            . $source_file_name
+            . $member_name
+            . $source_line_number;
+        
+        if (!empty($title)) {
+            if (is_string($title)) {
+                $string .= $title . $tab;
+            } else {
+                if($this->isTestModeEnabled()) {
+                    $string .= "\r\n" . json_encode($title, JSON_PRETTY_PRINT) . "\r\n";
+                } else{
+                    $string .= json_encode($title) . $tab;
+                }
+            }
+        }
+
+        $string .= $d . "\r\n\r\n";
+        
+        try {
+            switch ($this->config->isDebugEnabled(true)) {
+                case 3: // save log file per days
+                    $log_file_name = 'Nuvei-' . date('Y-m-d');
+                    break;
+                
+                case 2: // save single log file
+                    $log_file_name = 'Nuvei';
+                    break;
+                
+                case 1: // save both files
+                    $log_file_name = 'Nuvei';
+                    
+                    $this->saveFile($logsPath, date('Y-m-d') . '.log', $string, FILE_APPEND);
+                    break;
+                
+                default:
+                    return;
+            }
+            
+            return $this->saveFile($logsPath, $log_file_name . '.log', $string, FILE_APPEND);
+        } catch (exception $e) {
+            return;
+        }
+    }
+    
+    /**
+     * A single place to save files.
+     * 
+     * @param string $path
+     * @param mixed $data
+     * @param string $name The file name with extension. Append it to the $path.
+     * @param int $option A PHP constant like FILE_APPEND.
+     * 
+     * @return bool
+     */
+    public function saveFile($path, $name, $data, $option)
+    {
+        if(is_object($this->fileSystem) && $this->fileSystem->isDirectory($path)) {
+            return $this->fileSystem->filePutContents(
+                $path . DIRECTORY_SEPARATOR . $name,
+                $data,
+                $option
+            );
+        }
+        
+        if(is_dir($path)) {
+            return file_put_contents($path . DIRECTORY_SEPARATOR . $name, $data, $option);
+        }
+        
+        return false;
+    }
+}