+++ Model/Request/CreatePlan.php
@@ -8,7 +8,6 @@ use Nuvei\Checkout\Model\RequestInterface;
 class CreatePlan extends AbstractRequest implements RequestInterface
 {
     protected $requestFactory;
-    protected $config;
     protected $storeManager;
     
     public function __construct(
@@ -17,13 +16,14 @@ class CreatePlan extends AbstractRequest implements RequestInterface
         \Nuvei\Checkout\Lib\Http\Client\Curl $curl,
         \Nuvei\Checkout\Model\Response\Factory $responseFactory,
         \Nuvei\Checkout\Model\Request\Factory $requestFactory,
-        \Magento\Store\Model\StoreManagerInterface $storeManager
+        \Magento\Store\Model\StoreManagerInterface $storeManager,
+        \Nuvei\Checkout\Model\ReaderWriter $readerWriter
     ) {
         parent::__construct(
-            $logger,
             $config,
             $curl,
-            $responseFactory
+            $responseFactory,
+            $readerWriter
         );
 
         $this->requestFactory   = $requestFactory;