+++ Model/Request/GetPlansList.php
@@ -13,18 +13,18 @@ class GetPlansList extends AbstractRequest implements RequestInterface
     private $fileSystem;
     
     public function __construct(
-        \Nuvei\Checkout\Model\Logger $logger,
         \Nuvei\Checkout\Model\Config $config,
         \Nuvei\Checkout\Lib\Http\Client\Curl $curl,
         \Nuvei\Checkout\Model\Response\Factory $responseFactory,
         \Nuvei\Checkout\Model\Request\Factory $requestFactory,
-        \Magento\Framework\Filesystem\DriverInterface $fileSystem
+        \Magento\Framework\Filesystem\DriverInterface $fileSystem,
+        \Nuvei\Checkout\Model\ReaderWriter $readerWriter
     ) {
         parent::__construct(
-            $logger,
             $config,
             $curl,
-            $responseFactory
+            $responseFactory,
+            $readerWriter
         );
 
         $this->requestFactory   = $requestFactory;