+++ Model/Request/GetSessionToken.php
@@ -8,23 +8,22 @@ use Nuvei\Checkout\Model\RequestInterface;
 class GetSessionToken extends AbstractRequest implements RequestInterface
 {
     protected $requestFactory;
-    protected $config;
     
     public function __construct(
-        \Nuvei\Checkout\Model\Logger $logger,
         \Nuvei\Checkout\Model\Config $config,
         \Nuvei\Checkout\Lib\Http\Client\Curl $curl,
         \Nuvei\Checkout\Model\Response\Factory $responseFactory,
-        \Nuvei\Checkout\Model\Request\Factory $requestFactory
+        \Nuvei\Checkout\Model\Request\Factory $requestFactory,
+        \Nuvei\Checkout\Model\ReaderWriter $readerWriter
     ) {
         parent::__construct(
-            $logger,
             $config,
             $curl,
-            $responseFactory
+            $responseFactory,
+            $readerWriter
         );
 
-        $this->requestFactory   = $requestFactory;
+        $this->requestFactory = $requestFactory;
     }
     
     public function process()